+++ Sources/Yams/YamlError.swift
@@ -80,13 +80,6 @@ public enum YamlError: Error {
             return text + " in line \(mark.line), column \(mark.column)\n"
         }
     }
-
-#if swift(>=4.1.50)
-    @available(*, unavailable, renamed: "reader(problem:offset:value:yaml:)")
-    public static func reader(problem: String, byteOffset: Int, value: Int32, yaml: String) {
-        fatalError("unavailable")
-    }
-#endif
 }
 
 extension YamlError {